Implement visitor pattern
Since this is a sealed hierarchy, I think I could use the new instanceof
pattern matching to avoid the need for the visitor pattern. But it seems
like that won't be as efficient as visitors because it'll take O(n)
checks to find the right branch to execute as opposed to the O(1) for
the visitor polymorphic method dispatch.

It seems like extending switch with the new instanceof pattern matching
is in the offing, and that could probably be optimized to be O(1).